[2270] Move modals from sirius-components to sirius-web
Bug: https://github.com/eclipse-sirius/sirius-components/issues/2270
