+++ packages/core/frontend/sirius-components/src/index.ts
@@ -15,9 +15,4 @@ export * from './core/contextmenu/ContextMenu';
 export * from './core/form/Form';
 export * from './errors/ErrorBoundary';
 export * from './icons';
-export * from './modals/delete-project/DeleteProjectModal';
-export * from './modals/new-object/NewObjectModal';
-export * from './modals/new-representation/NewRepresentationModal';
-export * from './modals/new-root-object/NewRootObjectModal';
-export * from './modals/rename-project/RenameProjectModal';
 export * from './views/FormContainer';
