+++ packages/sirius-web/frontend/sirius-web/src/modals/new-object/NewObjectModal.types.ts
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2021, 2022 Obeo.
+ * Copyright (c) 2021, 2023 Obeo.
  * This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v2.0
  * which accompanies this distribution, and is available at
@@ -10,18 +10,14 @@
  * Contributors:
  *     Obeo - initial API and implementation
  *******************************************************************************/
-import { Selection } from '../../workbench/Workbench.types';
 
-export interface NewRootObjectModalProps {
+import { Selection } from '@eclipse-sirius/sirius-components-core';
+
+export interface NewObjectModalProps {
   editingContextId: string;
   item: any;
   onObjectCreated: (object: Selection) => void;
   onClose: () => void;
-}
-
-export interface Domain {
-  id: string;
-  label: string;
 }
 
 export interface ChildCreationDescription {
@@ -29,21 +25,12 @@ export interface ChildCreationDescription {
   label: string;
 }
 
-export interface GQLGetDomainsQueryVariables {
+export interface GQLGetChildCreationDescriptionsQueryVariables {
   editingContextId: string;
+  kind: string;
 }
 
-export interface GQLGetDomainsQueryData {
-  viewer: GQLViewer;
-}
-
-export interface GQLGetRootObjectCreationDescriptionsQueryVariables {
-  editingContextId: string;
-  domainId: string;
-  suggested: boolean;
-}
-
-export interface GQLGetRootObjectCreationDescriptionsQueryData {
+export interface GQLGetChildCreationDescriptionsQueryData {
   viewer: GQLViewer;
 }
 
@@ -52,8 +39,7 @@ export interface GQLViewer {
 }
 
 export interface GQLEditingContext {
-  rootObjectCreationDescriptions: GQLChildCreationDescription[];
-  domains: GQLDomain[];
+  childCreationDescriptions: GQLChildCreationDescription[];
 }
 
 export interface GQLChildCreationDescription {
@@ -61,20 +47,15 @@ export interface GQLChildCreationDescription {
   label: string;
 }
 
-export interface GQLDomain {
-  id: string;
-  label: string;
+export interface GQLCreateChildMutationData {
+  createChild: GQLCreateChildPayload;
 }
 
-export interface GQLCreateRootObjectMutationData {
-  createRootObject: GQLCreateRootObjectPayload;
-}
-
-export interface GQLCreateRootObjectPayload {
+export interface GQLCreateChildPayload {
   __typename: string;
 }
 
-export interface GQLCreateRootObjectSuccessPayload extends GQLCreateRootObjectPayload {
+export interface GQLCreateChildSuccessPayload extends GQLCreateChildPayload {
   id: string;
   object: GQLObject;
 }
@@ -85,6 +66,6 @@ export interface GQLObject {
   kind: string;
 }
 
-export interface GQLErrorPayload extends GQLCreateRootObjectPayload {
+export interface GQLErrorPayload extends GQLCreateChildPayload {
   message: string;
 }
