+++ packages/sirius-web/frontend/sirius-web/src/modals/new-representation/NewRepresentationModal.types.ts
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2021, 2022 Obeo.
+ * Copyright (c) 2021, 2023 Obeo.
  * This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v2.0
  * which accompanies this distribution, and is available at
@@ -10,7 +10,8 @@
  * Contributors:
  *     Obeo - initial API and implementation
  *******************************************************************************/
-import { Selection } from '../../workbench/Workbench.types';
+import { Selection } from '@eclipse-sirius/sirius-components-core';
+
 export interface NewRepresentationModalProps {
   editingContextId: string;
   item: any;
