+++ packages/sirius-web/frontend/sirius-web/src/modals/new-representation/NewRepresentationModalMachine.ts
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2021, 2022 Obeo.
+ * Copyright (c) 2021, 2023 Obeo.
  * This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v2.0
  * which accompanies this distribution, and is available at
@@ -16,7 +16,7 @@ import {
   GQLGetRepresentationDescriptionsQueryData,
   GQLRepresentationDescriptionMetadata,
 } from 'modals/new-representation/NewRepresentationModal.types';
-import { assign, Machine } from 'xstate';
+import { Machine, assign } from 'xstate';
 
 export interface NewRepresentationModalStateSchema {
   states: {
