+++ packages/sirius-web/frontend/sirius-web/src/modals/new-root-object/NewRootObjectModal.types.ts
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2021, 2022 Obeo.
+ * Copyright (c) 2021, 2023 Obeo.
  * This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v2.0
  * which accompanies this distribution, and is available at
@@ -10,13 +10,18 @@
  * Contributors:
  *     Obeo - initial API and implementation
  *******************************************************************************/
-import { Selection } from '../../workbench/Workbench.types';
+import { Selection } from '@eclipse-sirius/sirius-components-core';
 
-export interface NewObjectModalProps {
+export interface NewRootObjectModalProps {
   editingContextId: string;
   item: any;
   onObjectCreated: (object: Selection) => void;
   onClose: () => void;
+}
+
+export interface Domain {
+  id: string;
+  label: string;
 }
 
 export interface ChildCreationDescription {
@@ -24,12 +29,21 @@ export interface ChildCreationDescription {
   label: string;
 }
 
-export interface GQLGetChildCreationDescriptionsQueryVariables {
+export interface GQLGetDomainsQueryVariables {
   editingContextId: string;
-  kind: string;
 }
 
-export interface GQLGetChildCreationDescriptionsQueryData {
+export interface GQLGetDomainsQueryData {
+  viewer: GQLViewer;
+}
+
+export interface GQLGetRootObjectCreationDescriptionsQueryVariables {
+  editingContextId: string;
+  domainId: string;
+  suggested: boolean;
+}
+
+export interface GQLGetRootObjectCreationDescriptionsQueryData {
   viewer: GQLViewer;
 }
 
@@ -38,7 +52,8 @@ export interface GQLViewer {
 }
 
 export interface GQLEditingContext {
-  childCreationDescriptions: GQLChildCreationDescription[];
+  rootObjectCreationDescriptions: GQLChildCreationDescription[];
+  domains: GQLDomain[];
 }
 
 export interface GQLChildCreationDescription {
@@ -46,15 +61,20 @@ export interface GQLChildCreationDescription {
   label: string;
 }
 
-export interface GQLCreateChildMutationData {
-  createChild: GQLCreateChildPayload;
+export interface GQLDomain {
+  id: string;
+  label: string;
 }
 
-export interface GQLCreateChildPayload {
+export interface GQLCreateRootObjectMutationData {
+  createRootObject: GQLCreateRootObjectPayload;
+}
+
+export interface GQLCreateRootObjectPayload {
   __typename: string;
 }
 
-export interface GQLCreateChildSuccessPayload extends GQLCreateChildPayload {
+export interface GQLCreateRootObjectSuccessPayload extends GQLCreateRootObjectPayload {
   id: string;
   object: GQLObject;
 }
@@ -65,6 +85,6 @@ export interface GQLObject {
   kind: string;
 }
 
-export interface GQLErrorPayload extends GQLCreateChildPayload {
+export interface GQLErrorPayload extends GQLCreateRootObjectPayload {
   message: string;
 }
