+++ packages/sirius-web/frontend/sirius-web/src/modals/rename-project/RenameProjectModal.tsx
@@ -19,7 +19,7 @@ import DialogContent from '@material-ui/core/DialogContent';
 import DialogTitle from '@material-ui/core/DialogTitle';
 import TextField from '@material-ui/core/TextField';
 import { useMachine } from '@xstate/react';
-import React, { useEffect } from 'react';
+import { useEffect } from 'react';
 import {
   GQLErrorPayload,
   GQLRenameProjectMutationData,
