+++ packages/sirius-web/frontend/sirius-web/src/modals/rename-project/RenameProjectModalMachine.ts
