+++ packages/sirius-web/frontend/sirius-web/src/views/edit-project/ObjectTreeItemContextMenuContribution.tsx
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2021, 2022 Obeo.
+ * Copyright (c) 2021, 2023 Obeo.
  * This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v2.0
  * which accompanies this distribution, and is available at
@@ -10,7 +10,6 @@
  * Contributors:
  *     Obeo - initial API and implementation
  *******************************************************************************/
-import { NewObjectModal, NewRepresentationModal } from '@eclipse-sirius/sirius-components';
 import { Selection } from '@eclipse-sirius/sirius-components-core';
 import { TreeItemContextMenuComponentProps } from '@eclipse-sirius/sirius-components-trees';
 import ListItemIcon from '@material-ui/core/ListItemIcon';
@@ -18,6 +17,8 @@ import ListItemText from '@material-ui/core/ListItemText';
 import MenuItem from '@material-ui/core/MenuItem';
 import AddIcon from '@material-ui/icons/Add';
 import { forwardRef, Fragment, useState } from 'react';
+import { NewObjectModal } from '../../modals/new-object/NewObjectModal';
+import { NewRepresentationModal } from '../../modals/new-representation/NewRepresentationModal';
 
 type Modal = 'CreateNewObject' | 'CreateNewRepresentation';
